Paired with Colin, who works mostly with web techs
We started setting up testing in php, but later we decided that
Colin would be more comfortable using JavaScript, so we set up some
mocha tests. The code for testing the amount of live neighbours for any
given cell had to be put on hold with extensive examples on pen and
paper of Conway's Game of Life.