Session 03 - Pride
Pride is very comfortable with using his IDE of choice - WebStorm - we
were able to exchange some tips with using the ide. He had an old
version of node installed on his machine, so we couldn't use ES6 as
easily as we would have like. The time spent in the previous session on
setting up mocha allowed us to quickly start testing.